Next step for DML

+++ src/S031.MetaStack.Core/ORM/JMXFactory/Interfaces/IJMXProvider.cs
@@ -11,17 +11,17 @@ namespace S031.MetaStack.WinForms.ORM
 {
 	public interface IJMXProvider
     {
-		JMXObject Read(JMXObjectName objectName, int id);
+		JMXObject Read(JMXObjectName objectName, long id);
 		JMXObject Read(JMXObjectName objectName, string alterKeyIndexName, params object[] parameters);
 		void Delete(JMXObject jmxObject);
-		void Delete(JMXObjectName objectName, int id);
-		int Save(JMXObject jmxObject, bool isNew);
-		int SetState(string objectName, int handle, int newState, IDictionary<string, object> paramList);
-		Task<JMXObject> ReadAsync(JMXObjectName objectName, int id);
+		void Delete(JMXObjectName objectName, long id);
+		long Save(JMXObject jmxObject, bool isNew);
+		int SetState(string objectName, long handle, int newState, IDictionary<string, object> paramList);
+		Task<JMXObject> ReadAsync(JMXObjectName objectName, long id);
 		Task<JMXObject> ReadAsync(JMXObjectName objectName, string alternameIndexName, params object[] parameters);
 		Task DeleteAsync(JMXObject jmxObject);
-		Task DeleteAsync(JMXObjectName objectName, int id);
-		Task<int> SaveAsync(JMXObject jmxObject, bool isNew);
-		Task<int> SetStateAsync(string objectName, int handle, int newState, IDictionary<string, object> paramList);
+		Task DeleteAsync(JMXObjectName objectName, long id);
+		Task<long> SaveAsync(JMXObject jmxObject, bool isNew);
+		Task<int> SetStateAsync(string objectName, long handle, int newState, IDictionary<string, object> paramList);
 	}
 }
